Move _barycentric_coordinates function to uxgrid
This function is now only needed within the uxgrid module
